+++ app/src/main/java/com/github/xzwj87/todolist/schedule/ui/activity/ScheduleDetailActivity.java
@@ -40,9 +40,9 @@ public class ScheduleDetailActivity extends AppCompatActivity {
         if (savedInstanceState == null) {
             Bundle arguments = new Bundle();
             Log.v(LOG_TAG, "onCreate(): intent = " + getIntent().getData());
-            int id = getIntent().getIntExtra(SCHEDULE_ID, 0);
+            long id = getIntent().getLongExtra(SCHEDULE_ID, 0);
             Log.v(LOG_TAG, "onCreate(): id = " + id);
-            arguments.putInt(ScheduleDetailFragment.SCHEDULE_ID, id);
+            arguments.putLong(ScheduleDetailFragment.SCHEDULE_ID, id);
 
             ScheduleDetailFragment fragment = new ScheduleDetailFragment();
             fragment.setArguments(arguments);
